+++ src/FatGraphs.jl
@@ -29,7 +29,7 @@ Edge, is_ordered, reverse,
 # core
 vertices, edges, src, dst,
 graph, digraph,
-fadj, badj, adj,
+fadj, badj, adjlist,
 in_edges, out_edges, has_vertex, has_edge, is_directed,
 nv, ne, add_edge!, rem_edge!, add_vertex!, add_vertices!,
 indegree, outdegree, degree, degree_histogram, density, Δ, δ,