+++ test/core/core.jl
@@ -40,8 +40,8 @@ end
 @test indegree(h,1) == 0
 @test outdegree(h) == [3, 1, 1, 0, 0]
 @test outdegree(h,1) == 3
-@test in_neighbors(h,5) == FatGraphs.badj(h)[5] == FatGraphs.badj(h,5) == [2, 3]
-@test out_neighbors(h,1) == FatGraphs.fadj(h)[1] == FatGraphs.fadj(h,1) == [2, 3, 4]
+@test in_neighbors(h,5) == badj(h)[5]  == [2, 3]
+@test out_neighbors(h,1) == fadj(h)[1]  == [2, 3, 4]
 
 @test p1 == g2
 @test issubset(h2, h1)
@@ -194,9 +194,9 @@ h = DiGraph(5)
 @test add_edge!(h, e4)
 @test add_edge!(h, e5)
 
-@test fadj(g)[1] == fadj(g,1) ==
-    badj(g)[1] == badj(g,1) ==
-    adj(g)[1] == adj(g,1) == [2,3,4]
+@test fadj(g)[1] == out_neighbors(g,1) ==
+    badj(g)[1] == in_neighbors(g,1) ==
+    adjlist(g)[1] == [2,3,4]
 
 @test sprint(show, h4) == "{7, 0} directed graph"
 @test sprint(show, h5) == "empty directed graph"